fb_id converted in bignum

+++ db/migrate/20120710230702_create_users.rb
@@ -5,7 +5,7 @@ class CreateUsers < ActiveRecord::Migration
       t.boolean :gender
       t.string :username
       t.string :ip
-      t.integer :fb_id
+      t.bigint :fb_id
       t.string :fb_first_name
       t.string :fb_last_name
       t.string :location
